requirements.txt
Flask==2.1.2
Flask_SQLAlchemy==2.5.1
Jinja2==3.1.2
psycopg2_binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.36
